Moving components to 'components' folder
