+++ win-linux/src/cascapplicationmanagerwrapper.cpp
@@ -17,13 +17,13 @@
 
 #include "cstyletweaks.h"
 #include "defines.h"
-#include "cfiledialog.h"
+#include "components/cfiledialog.h"
 #include "utils.h"
 #include "common/Types.h"
 #include "common/File.h"
 #include "ctabundockevent.h"
 #include "clangater.h"
-#include "cmessage.h"
+#include "components/cmessage.h"
 #include "ceditortools.h"
 #include "cfilechecker.h"
 #include "OfficeFileFormats.h"
