+++ win-linux/src/components/asctabwidget.h
@@ -132,7 +132,6 @@ private:
 
 signals:
 //    void sendAddEditor();
-    void closeAppRequest();
     void editorInserted(int, int);
     void editorRemoved(int, int);
 
@@ -157,7 +156,7 @@ public:
     int  count(const std::wstring& portal, bool exclude = false);
     bool hasForPortal(const QString&);
 
-    void updateScaling(double) override;
+    void updateScalingFactor(double) override;
 protected:
     void tabInserted(int index) override;
     void tabRemoved(int index) override;
