+++ win-linux/src/components/canimatedicon.h
