+++ win-linux/src/components/cdownloadwidget.cpp
@@ -30,7 +30,7 @@
  *
 */
 
-#include "cdownloadwidget.h"
+#include "components/cdownloadwidget.h"
 
 #include <QVBoxLayout>
 #include <QLabel>
@@ -43,7 +43,7 @@
 #include <QWidgetAction>
 #include "qcefview.h"
 #include "common/Types.h"
-#include "cwindowbase.h"
+#include "windows/cwindowbase.h"
 
 #include <QDebug>
 
@@ -345,9 +345,9 @@ void CDownloadWidget::applyScaling(double factor)
     QPixmap::grabWidget(m_pToolButton->menu());
 }
 
-void CDownloadWidget::updateScaling(double f)
+void CDownloadWidget::updateScalingFactor(double f)
 {
-    CScalingWrapper::updateScaling(f);
+    CScalingWrapper::updateScalingFactor(f);
     applyScaling(scaling());
 }
